Update FZF default command to use fd
Prior to this change, FZF was configured to use fd, but this was done in
the zshenv file, and it was not done very effectively. This change moves
the default command into .zshrc where it belongs. Furthermore, the
command using fd is smarter, looking for files and directories,
following sym links but avoiding .git files. To tweak which directories
and files should be ignored, use the global fd ignore file
(~/.config/fd/ignore).