Refactor zshrc
Prior to this change, my zshrc was all in one file. This worked but made
it harder to find things and to keep everything organized. This refactor
makes it easier to scale and maintain my zshrc.

Furthermore, it introduces a way to add local overrides to each of the
files. These local overrides are ignored by git. For the zshenv the
override should be $ZDOTDIR/zshenv.local.zsh. For all other local
overrides, they should follow the pattern $XDOTDIR/${section_name}.zsh
-> $XDOTDIR/${section_name}.local.zsh