Use fzf-tab for completion
This hooks fzf-tab into the zsh completion system. I like how this works
for most things, but I do not like how it completes files and
directories. I think completing at each level of the path is too slow.
A better way, which I will look into in the future is to use fzf to
fuzzy find the complete path to the directory or path.