Remove local overrides check
Prior to this change, the zshrc looked for a local overrides file and
applied the changes after running the zshrc. This functionality was
already partially removed in an earlier change, but this gets rid of the
vestigial code. Local overrides are better handled using a local code
branch than explicit override files.